Removed Data project, and moved data files out of solution
Removed Data project, and moved data files out of solution
* Removed Data Project
* Data files are now in /data/data-files
* Moved JsonDataFile into CarbonAware core project
* Created Data namespace for data and file classes
* Removed data.sample namespace
* Added file copy to CLI project as post build event
* Added file copy to WebApi project as post build event
