+++ src/dotnet/CarbonAware.Plugins.BasicJsonPlugin.Tests/CarbonAwareLogicPluginSampleTest.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using CarbonAware.Data;
 using Microsoft.Extensions.Configuration;
 using NUnit.Framework;
