+++ src/dotnet/CarbonAware.Tools.AWSDatacenterDummyEmissionsGenerator/AwsRegionStaticDataService.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using CarbonAware.Data;
 using Newtonsoft.Json;
 
 namespace CarbonAware.Tools
