+++ src/dotnet/CarbonAware.Tools.AzureDatacenterDummyEmissionsGenerator/AzureRegionDummyDataGenerator.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using CarbonAware;
+using CarbonAware.Data;
 using Newtonsoft.Json;
 
 namespace CarbonAware.Tools
