+++ src/dotnet/CarbonAware/CarbonAwareCore.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using CarbonAware.Data;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
