+++ src/dotnet/CarbonAware/Data/EmissionsData.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CarbonAware
+namespace CarbonAware.Data
 {
     [Serializable]
     public class EmissionsData
