+++ src/dotnet/CarbonAwareSDK.sln
@@ -15,8 +15,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CarbonAware.Tools.AzureData
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CarbonAware.Tools.AWSDatacenterDummyEmissionsGenerator", "CarbonAware.Tools.AWSDatacenterDummyEmissionsGenerator\CarbonAware.Tools.AWSDatacenterDummyEmissionsGenerator.csproj", "{16B15877-9DFC-43E8-84FE-BA67C4C71899}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CarbonAware.Data.Sample", "CarbonAware.Data.Sample\CarbonAware.Data.Sample.csproj", "{34B58DE8-6327-4BDF-8D58-7E76146CA85C}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CarbonAware.WebApi", "CarbonAware.WebApi\CarbonAware.WebApi.csproj", "{1A606EAE-C8BE-454F-B3F7-C649C9293B65}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarbonAware.Tests", "CarbonAware.Tests\CarbonAware.Tests.csproj", "{834D7E6B-8807-4808-96CF-1BCA36C17F6D}"
@@ -51,10 +49,6 @@ Global
 		{16B15877-9DFC-43E8-84FE-BA67C4C71899}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{16B15877-9DFC-43E8-84FE-BA67C4C71899}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{16B15877-9DFC-43E8-84FE-BA67C4C71899}.Release|Any CPU.Build.0 = Release|Any CPU
-		{34B58DE8-6327-4BDF-8D58-7E76146CA85C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{34B58DE8-6327-4BDF-8D58-7E76146CA85C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{34B58DE8-6327-4BDF-8D58-7E76146CA85C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{34B58DE8-6327-4BDF-8D58-7E76146CA85C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1A606EAE-C8BE-454F-B3F7-C649C9293B65}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1A606EAE-C8BE-454F-B3F7-C649C9293B65}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1A606EAE-C8BE-454F-B3F7-C649C9293B65}.Release|Any CPU.ActiveCfg = Release|Any CPU
