+++ src/dotnet/data/data-files/aws-regions.json
