+++ src/dotnet/data/data-files/dummy-data-aws-emissions.json
